requirements.txt
fastapi
uvicorn
pillow
pygame
python-multipart